Separate backup and restore cluster functions
We separate the backup specific and restore specific functions that make
ssh calls into new backup/remote.go and restore/remote.go files.

Also, separate compression from cluster file into compression.go

Author: Karen Huddleston <khuddleston@pivotal.io>
Author: Chris Hajas <chajas@pivotal.io>
