+++ integration/integration_suite_test.go
@@ -92,7 +92,7 @@ var _ = AfterSuite(func() {
 })
 
 func setupTestFilespace(cluster utils.Cluster) {
-	cluster.CreateBackupDirectoriesOnAllHosts()
+	backup.CreateBackupDirectoriesOnAllHosts(cluster)
 	// Construct a filespace config like the one that gpfilespace generates
 	filespaceConfigQuery := `COPY (SELECT hostname || ':' || dbid || ':/tmp/test_dir/' || preferred_role || content FROM gp_segment_configuration AS subselect) TO '/tmp/temp_filespace_config';`
 	testutils.AssertQueryRuns(connection, filespaceConfigQuery)
